% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_geo.R
\name{om_geo}
\alias{om_geo}
\title{Load OnMarg data}
\usage{
om_geo(year, level)
}
\arguments{
\item{year}{Integer year of data to load}

\item{level}{The level of precision to load, this can be "DAUID", "CTUID", "CSDUID", "CCSUID", "CDUID", "CMAUID", "PHUUID", "LHINUID", or "LHIN_SRUID"}
}
\description{
This is a function for loading OnMarg data into an sf object for spatial analysis.
}
\details{
If a year or level is used that does not exist or is not implemented, an error message will be produced.
}
\examples{
DA_2016 <- om_geo(2016, "DAUID")
DA_2011 <- om_geo(2011, "DAUID")
CSD_2016 <- om_geo(2016, "CSDUID")
CSD_2011 <- om_geo(2011, "CSDUID")
}
