% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_quint.R
\name{om_quint}
\alias{om_quint}
\title{This function converts an arbitrary vector of values into corresponding quintile scores.}
\usage{
om_quint(x)
}
\arguments{
\item{x}{Vector of values to recalculate quintiles for}
}
\value{
Vector of quintile scores for each element in the input vector
}
\description{
NA values are ignored and left NA
}
\details{
It can be used to recalculate the quintile scores for subsets of the OnMarg dataset.
}
\examples{
\dontrun{
city_data$DEPRIVATION_Q_DA16 <- om_quint(city_data$DEPRIVATION_DA16)
}
}
